Added changes for Azure policy task
