--- Tasks/AzurePolicyTaskV1/Tests/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "vsts-tasks-azurewebappslotswap",
-  "version": "0.1.0",
-  "description": "VSTS Azure WebApp Slot Swap",
-  "main": "azurewebappslotswap.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/Microsoft/vsts-tasks.git"
-  },
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Microsoft/vsts-tasks/issues"
-  },
-  "homepage": "https://github.com/Microsoft/vsts-tasks#readme",
-  "devDependencies": {
-    "nock": "9.0.11"
-  }
-}
